refactor: extract SQLAlchemy models to standalone module
Alembic's env.py imported Base from vectordb/utils.py, which transitively
pulled in the entire app: utils.py -> logger -> config -> Hydra, and
via Python's package init chain: components.indexer.__init__ -> indexer ->
chunker -> prompts -> file I/O. This meant `alembic upgrade head` failed
unless run from the exact working directory with all config and prompt
files accessible — making it fragile in Docker and CI contexts.

Extract the pure data definitions (Base, File, Partition, User,
PartitionMembership) into a new vectordb/models.py that only depends on
sqlalchemy and datetime. Alembic's env.py now imports Base from models.py
instead of utils.py, and utils.py re-imports the models so all existing
code continues to work unchanged.
